requirements.txt
Django
daphne
celery
SQLAlchemy
djangorestframework
aiohttp